@@ -5,6 +5,7 @@
   "module": "fumifier.js",
   "main": "fumifier.js",
   "typings": "fumifier.d.ts",
+  "type": "module",
   "scripts": {
     "fume-test": "node src/fume.test.js",
     "pretest": "npm run lint",
@@ -35,18 +36,19 @@
   ],
   "devDependencies": {
     "@babel/cli": "^7.27.2",
-    "@babel/core": "^7.27.4",
+    "@babel/core": "^7.27.7",
     "@babel/preset-env": "^7.27.2",
-    "@outburn/structure-navigator": "^1.0.2",
+    "@outburn/structure-navigator": "^1.2.0",
     "browserify": "^16.5.2",
     "chai": "^4.5.0",
     "chai-as-promised": "^7.1.2",
     "eslint": "8.0.0",
     "eslint-plugin-ideal": "^0.1.3",
     "eslint-plugin-promise": "^6.6.0",
+    "fhir-snapshot-generator": "^1.3.0",
     "jsdoc": "^4.0.4",
     "mkdirp": "^1.0.4",
-    "mocha": "^11.7.0",
+    "mocha": "^11.7.1",
     "mocha-lcov-reporter": "^1.3.0",
     "nyc": "^15.1.0",
     "regenerator": "^0.14.12",
@@ -54,6 +56,6 @@
     "uglify-es": "^3.3.10"
   },
   "engines": {
-    "node": ">= 8"
+    "node": ">= 20"
   }
 }
